FORAY_DJANGO/requirements-dev.txt
# Development dependencies for testing and code quality
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
faker==19.6.2

# Code quality tools
black==23.7.0
flake8==6.0.0
isort==5.12.0

# Security scanning
bandit==1.7.5

# Environment management
python-dotenv==1.0.0

# Database support
psycopg2-binary==2.9.7

# Production server
gunicorn==21.2.0